Update to support new parameters and fix some issues. (#3)
* Update to support new parameters and fix some issues.

* Fixed errors in test and suppression rule interfaces

* Remove esModuleInterop option. Fix issue with renamed interfaces.

* Add check for accessToken or callback.

* Add check for undefined accessToken or callback and fixed lint issues.

* Refactor redundant check for accessToken

* Add imodel data extraction calls

* Remove webpack dependencies.

* Remove obsolete test config files

* Update to release version 1.0.0 and cleanup references and lint issues

* Dependency, exports and lint config cleanup

* Update package lock file

* Set release version to 0.2.0

+++ src/operations/suppressionRule/SuppressionRuleOperations.ts
@@ -2,10 +2,11 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
+import { OperationUtils } from "../OperationUtils";
 import { OperationsBase } from "../../base/OperationsBase";
 import { PreferReturn } from "../../base/interfaces/CommonInterfaces";
 import { EntityListIteratorImpl } from "../../base/iterators/EntityListIteratorImpl";
-import type { MinimalSuppressionRule, ResponseFromCreateSuppressionRule, ResponseFromGetSuppressionRule, ResponseFromGetSuppressionRuleList, ResponseFromGetSuppressionRuleListMinimal, ResponseFromUpdateSuppressionRule, SuppressionRule, SuppressionRuleDetails } from "../../base/interfaces/apiEntities/SuppressionRuleInterfaces";
+import type { MinimalSuppressionRule, ResponseFromCreateSuppressionRule, ResponseFromGetSuppressionRule, ResponseFromGetSuppressionRuleList, ResponseFromGetSuppressionRuleListMinimal, ResponseFromUpdateSuppressionRule, SuppressionRuleCreate, SuppressionRuleDetails, SuppressionRuleUpdate } from "../../base/interfaces/apiEntities/SuppressionRuleInterfaces";
 import type { EntityListIterator } from "../../base/iterators/EntityListIterator";
 import type { OperationOptions } from "../OperationOptions";
 import type { ParamsToCreateSuppressionRule, ParamsToDeleteSuppressionRule, ParamsToGetSuppressionRule, ParamsToGetSuppressionRuleList, ParamsToUpdateSuppressionRule } from "./SuppressionRuleOperationParams";
@@ -31,12 +32,14 @@ export class SuppressionRuleOperations<TOptions extends OperationOptions> extend
       const rules = (response as ResponseFromGetSuppressionRuleListMinimal).suppressionRules;
       return rules;
     };
-
+    OperationUtils.ensureAccessTokenProvided(params.accessToken, this._options.accessTokenCallback);
     return new EntityListIteratorImpl(async () => this.getEntityCollectionPage<MinimalSuppressionRule>({
+      // eslint-disable-next-line  @typescript-eslint/no-non-null-assertion
       accessToken: params.accessToken ?? await this._options.accessTokenCallback!(),
       url: this._options.urlFormatter.getRuleListUrl({urlParams: params.urlParams }),
       preferReturn: PreferReturn.Minimal,
       entityCollectionAccessor,
+      userMetadata: false,
     }));
   }
 
@@ -54,12 +57,14 @@ export class SuppressionRuleOperations<TOptions extends OperationOptions> extend
       const rules = (response as ResponseFromGetSuppressionRuleList).suppressionRules;
       return rules;
     };
-
+    OperationUtils.ensureAccessTokenProvided(params.accessToken, this._options.accessTokenCallback);
     return new EntityListIteratorImpl(async () => this.getEntityCollectionPage<SuppressionRuleDetails>({
+      // eslint-disable-next-line  @typescript-eslint/no-non-null-assertion
       accessToken: params.accessToken ?? await this._options.accessTokenCallback!(),
       url: this._options.urlFormatter.getRuleListUrl({urlParams: params.urlParams }),
       preferReturn: PreferReturn.Representation,
       entityCollectionAccessor,
+      userMetadata: params.userMetadata ?? false,
     }));
   }
 
@@ -71,10 +76,13 @@ export class SuppressionRuleOperations<TOptions extends OperationOptions> extend
    * @returns {Promise<SuppressionRuleDetails>} a Suppression Rule with specified id. See {@link SuppressionRuleDetails}.
    */
   public async getSingle(params: ParamsToGetSuppressionRule): Promise<SuppressionRuleDetails> {
-    const { accessToken, ruleId } = params;
+    const { accessToken, ruleId, userMetadata } = params;
+    OperationUtils.ensureAccessTokenProvided(accessToken, this._options.accessTokenCallback);
     const response = await this.sendGetRequest<ResponseFromGetSuppressionRule>({
+      // eslint-disable-next-line  @typescript-eslint/no-non-null-assertion
       accessToken: accessToken ?? await this._options.accessTokenCallback!(),
       url: this._options.urlFormatter.getSingleRuleUrl({ ruleId }),
+      userMetadata: userMetadata ?? false,
     });
     return response.suppressionRule;
   }
@@ -88,7 +96,9 @@ export class SuppressionRuleOperations<TOptions extends OperationOptions> extend
    */
   public async delete(params: ParamsToDeleteSuppressionRule): Promise<void> {
     const { accessToken, ruleId } = params;
+    OperationUtils.ensureAccessTokenProvided(accessToken, this._options.accessTokenCallback);
     await this.sendDeleteRequest<void>({
+      // eslint-disable-next-line  @typescript-eslint/no-non-null-assertion
       accessToken: accessToken ?? await this._options.accessTokenCallback!(),
       url: this._options.urlFormatter.deleteRuleUrl({ ruleId }),
     });
@@ -100,14 +110,16 @@ export class SuppressionRuleOperations<TOptions extends OperationOptions> extend
    * @param {ParamsToCreateSuppressionRule} params parameters for this operation. See {@link ParamsToCreateSuppressionRule}.
    * @returns {Promise<SuppressionRule>} newly created Suppression Rule. See {@link SuppressionRule}.
    */
-  public async create(params: ParamsToCreateSuppressionRule): Promise<SuppressionRule> {
+  public async create(params: ParamsToCreateSuppressionRule): Promise<SuppressionRuleCreate> {
     const body = {
       templateId: params.templateId,
       displayName: params.displayName,
       description: params.reason,
       parameters: params.parameters,
     };
+    OperationUtils.ensureAccessTokenProvided(params.accessToken, this._options.accessTokenCallback);
     const response = await this.sendPostRequest<ResponseFromCreateSuppressionRule>({
+      // eslint-disable-next-line  @typescript-eslint/no-non-null-assertion
       accessToken: params.accessToken ?? await this._options.accessTokenCallback!(),
       url: this._options.urlFormatter.createRuleUrl(),
       body,
@@ -122,12 +134,14 @@ export class SuppressionRuleOperations<TOptions extends OperationOptions> extend
    * @param {ParamsToUpdateSuppressionRule} params parameters for this operation. See {@link ParamsToUpdateSuppressionRule}.
    * @returns {Promise<SuppressionRule>} newly updated Suppression Rule. See {@link SuppressionRule}.
    */
-  public async update(params: ParamsToUpdateSuppressionRule): Promise<SuppressionRule> {
+  public async update(params: ParamsToUpdateSuppressionRule): Promise<SuppressionRuleUpdate> {
     const body = {
       displayName: params.displayName,
       description: params.reason,
     };
+    OperationUtils.ensureAccessTokenProvided(params.accessToken, this._options.accessTokenCallback);
     const response = await this.sendPutRequest<ResponseFromUpdateSuppressionRule>({
+      // eslint-disable-next-line  @typescript-eslint/no-non-null-assertion
       accessToken: params.accessToken ?? await this._options.accessTokenCallback!(),
       url: this._options.urlFormatter.updateRuleUrl(params),
       body,
