+++ src/operations/test/TestOperations.ts
@@ -2,6 +2,7 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
+import { OperationUtils } from "../OperationUtils";
 import { OperationsBase } from "../../base/OperationsBase";
 import { EntityListIteratorImpl } from "../../base/iterators/EntityListIteratorImpl";
 import { IModelsClient, NamedVersionOrderByProperty, OrderByOperator, toArray } from "@itwin/imodels-client-management";
@@ -33,11 +34,13 @@ export class TestOperations<TOptions extends OperationOptions> extends Operation
       const tests = (response as ResponseFromGetTestList).tests;
       return tests;
     };
-
+    OperationUtils.ensureAccessTokenProvided(params.accessToken, this._options.accessTokenCallback);
     return new EntityListIteratorImpl(async () => this.getEntityCollectionPage<TestItem>({
+      // eslint-disable-next-line  @typescript-eslint/no-non-null-assertion
       accessToken: params.accessToken ?? await this._options.accessTokenCallback!(),
       url: this._options.urlFormatter.getTestListUrl({ urlParams: params.urlParams }),
       entityCollectionAccessor,
+      userMetadata: params.userMetadata ?? false,
     }));
   }
 
@@ -49,10 +52,13 @@ export class TestOperations<TOptions extends OperationOptions> extends Operation
    * @returns {Promise<TestDetails>} a Test with specified id. See {@link TestDetails}.
    */
   public async getSingle(params: ParamsToGetTest): Promise<TestDetails> {
-    const { accessToken, testId } = params;
+    const { accessToken, testId, userMetadata } = params;
+    OperationUtils.ensureAccessTokenProvided(accessToken, this._options.accessTokenCallback);
     const response = await this.sendGetRequest<ResponseFromGetTest>({
+      // eslint-disable-next-line  @typescript-eslint/no-non-null-assertion
       accessToken: accessToken ?? await this._options.accessTokenCallback!(),
       url: this._options.urlFormatter.getSingleTestUrl({ testId }),
+      userMetadata: userMetadata ?? false,
     });
     return response.test;
   }
@@ -74,8 +80,11 @@ export class TestOperations<TOptions extends OperationOptions> extends Operation
       touchingTolerance: params.touchingTolerance,
       includeSubModels: params.includeSubModels,
       suppressionRules: params.suppressionRules,
+      advancedSettings: params.advancedSettings,
     };
+    OperationUtils.ensureAccessTokenProvided(params.accessToken, this._options.accessTokenCallback);
     const response = await this.sendPostRequest<ResponseFromCreateTest>({
+      // eslint-disable-next-line  @typescript-eslint/no-non-null-assertion
       accessToken: params.accessToken ?? await this._options.accessTokenCallback!(),
       url: this._options.urlFormatter.createTestUrl(),
       body,
@@ -100,8 +109,11 @@ export class TestOperations<TOptions extends OperationOptions> extends Operation
       touchingTolerance: params.touchingTolerance,
       includeSubModels: params.includeSubModels,
       suppressionRules: params.suppressionRules,
+      advancedSettings: params.advancedSettings,
     };
+    OperationUtils.ensureAccessTokenProvided(params.accessToken, this._options.accessTokenCallback);
     const response = await this.sendPutRequest<ResponseFromUpdateTest>({
+      // eslint-disable-next-line  @typescript-eslint/no-non-null-assertion
       accessToken: params.accessToken ?? await this._options.accessTokenCallback!(),
       url: this._options.urlFormatter.updateTestUrl(params),
       body,
@@ -117,17 +129,17 @@ export class TestOperations<TOptions extends OperationOptions> extends Operation
    * @returns {Promise<Run>} newly started Run. See {@link Run}.
    */
   public async runTest(params: ParamsToRunTest): Promise<Run | undefined> {
+    OperationUtils.ensureAccessTokenProvided(params.accessToken, this._options.accessTokenCallback);
     // If namedVersionId is not specified, then try to get latest version as default
     if (params.namedVersionId === undefined) {
       const iModelsClient: IModelsClient = new IModelsClient();
       let authorization: AuthorizationCallback;
       if (params.accessToken) {
         authorization = ClashDetectionClient.toAuthorizationCallback(params.accessToken);
-      } else if (this._options.accessTokenCallback) {
-        const accessToken = await this._options.accessTokenCallback();
-        authorization = ClashDetectionClient.toAuthorizationCallback(accessToken);
       } else {
-        throw new Error(`Access token is required`);
+        // eslint-disable-next-line  @typescript-eslint/no-non-null-assertion
+        const accessToken = await this._options.accessTokenCallback!();
+        authorization = ClashDetectionClient.toAuthorizationCallback(accessToken);
       }
       const getNamedVersionListParams: GetNamedVersionListParams = {
         authorization,
@@ -141,17 +153,19 @@ export class TestOperations<TOptions extends OperationOptions> extends Operation
       };
       const namedVersionsIterator: EntityListIterator<MinimalNamedVersion> = iModelsClient.namedVersions.getMinimalList(getNamedVersionListParams);
       const namedVersions: MinimalNamedVersion[] = await toArray(namedVersionsIterator);
-      if (namedVersions.length === 0)
+      if (namedVersions.length === 0) {
         return undefined;
-
+      }
       params.namedVersionId = namedVersions[0].id;
     }
     const body = {
       testId: params.testId,
       iModelId: params.iModelId,
       namedVersionId: params.namedVersionId,
+      testSettings: params.testSettings,
     };
     const response = await this.sendPostRequest<ResponseFromRunTest>({
+      // eslint-disable-next-line  @typescript-eslint/no-non-null-assertion
       accessToken: params.accessToken ?? await this._options.accessTokenCallback!(),
       url: this._options.urlFormatter.runTestUrl(),
       body,
@@ -169,7 +183,9 @@ export class TestOperations<TOptions extends OperationOptions> extends Operation
    */
   public async delete(params: ParamsToDeleteTest): Promise<void> {
     const { accessToken, testId } = params;
+    OperationUtils.ensureAccessTokenProvided(accessToken, this._options.accessTokenCallback);
     await this.sendDeleteRequest<void>({
+      // eslint-disable-next-line  @typescript-eslint/no-non-null-assertion
       accessToken: accessToken ?? await this._options.accessTokenCallback!(),
       url: this._options.urlFormatter.deleteTestUrl({ testId }),
     });
